pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openg2p-portal-api"
authors = [
  { name="OpenG2P", email="info@openg2p.org" },
]
description = ""
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
  "Operating System :: OS Independent",
]
dependencies = [
  "openg2p-fastapi-common",
  "openg2p-fastapi-auth",
  "boto3",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://openg2p.org"
Documentation = "https://docs.openg2p.org/"
Repository = "https://github.com/OpenG2P/openg2p-portal-api"
Source = "https://github.com/OpenG2P/openg2p-portal-api"

[tool.hatch.version]
path = "src/openg2p_portal_api/__init__.py"
